requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Deprecated==1.2.14
Django==4.2.3
django-oauth-toolkit==2.3.0
django-sslserver==0.22
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
